Move test to test_api.py
